gave components and product lists their own folders and orderd the productlist
